feat: refactor DataPermissionPolicy and PolicyType to improve namespace structure and enhance role/user associations
