+++ app/Model/Permission/Role.php
@@ -12,10 +12,12 @@ declare(strict_types=1);
 
 namespace App\Model\Permission;
 
+use App\Model\DataPermission\Policy;
 use App\Model\Enums\User\Status;
 use Carbon\Carbon;
 use Hyperf\Database\Model\Collection;
 use Hyperf\Database\Model\Events\Deleting;
+use Hyperf\Database\Model\Relations\BelongsTo;
 use Hyperf\Database\Model\Relations\BelongsToMany;
 use Hyperf\DbConnection\Model\Model as MineModel;
 
@@ -33,6 +35,7 @@ use Hyperf\DbConnection\Model\Model as MineModel;
  * @property string $remark 备注
  * @property Collection|Menu[] $menus
  * @property Collection|User[] $users
+ * @property Policy|null $policy 数据权限策略
  */
 final class Role extends MineModel
 {
@@ -89,5 +92,11 @@ final class Role extends MineModel
     {
         $this->users()->detach();
         $this->menus()->detach();
+    }
+
+
+    public function policy(): BelongsTo
+    {
+        return $this->belongsTo(Policy::class, 'policy_id', 'id');
     }
 }
